requirements.txt
fastapi
uvicorn
numpy
opencv_python
ultralytics
google-generativeai
